acerto navbar

+++ src/app/componentes/nav-bar/nav-bar.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from "@angular/router";
+import Swal from 'sweetalert2';
+import { Global } from '../../global';
 
 @Component({
   selector: 'app-nav-bar',
@@ -10,15 +12,26 @@ export class NavBarComponent implements OnInit {
 
   constructor(
     public router: Router,
+    public global: Global
   ) { }
 
   ngOnInit() {
   }
 
   sair = () => {
-    alert('logout')
-    localStorage.clear();
-    this.router.navigate(['/login']);
+    Swal.fire({
+      icon:'warning',
+      title:'Atenção',
+      text:'Deseja mesmo sair?',
+      showCancelButton: true
+    }).then(res => {
+      if(res.isConfirmed){
+        localStorage.clear();
+        this.router.navigate(['/login']);
+        this.global.nav = false;
+      }
+    })
   }
+  
 
 }
